chore: move arangoValidationHelper to separate file, import earlier
