+++ js/apps/system/_admin/aardvark/APP/react/src/App.js
@@ -8,7 +8,7 @@ require('nvd3');
 
 // parse prometheus
 const parsePrometheusTextFormat = require('parse-prometheus-text-format');
-
+require('../../frontend/js/arango/arangoValidationHelper.js');
 // import new react views
 // require('./views/shards/ShardsReactView');
 require('./views/analyzers/AnalyzersReactView');
